Refactor crown files
