+++ src/assets/scss/index.scss
@@ -17,12 +17,10 @@
 @import './components/components.modeler';
 @import './components/components.toolbar';
 @import './components/components.inspector';
-@import './components/components.controls';
 @import './components/components.sequenceFlowFormSelect';
 @import './components/components.statusBar';
 @import './components/components.miniPaper';
 @import './components/components.validationStatus';
-@import './components/components.crownDropdown';
 
 // Utilities – utilities and helper classes with ability to override anything which goes before in the triangle, eg. hide helper class
 @import './utilities/utilities.cursor';
