+++ src/components/crown/crownButtons/deleteButton.vue
@@ -20,7 +20,7 @@
 
 <script>
 import trashIcon from '@/assets/trash-alt-solid.svg';
-import CrownButton from '@/components/crownButton';
+import CrownButton from '@/components/crown/crownButtons/crownButton';
 
 export default {
   components: { CrownButton },
