+++ src/components/crown/crownButtons/sequenceFlowButton.vue
@@ -12,7 +12,7 @@
 </template>
 <script>
 import sequenceFlow from '@/assets/connect-elements.svg';
-import CrownButton from '@/components/crownButton';
+import CrownButton from '@/components/crown/crownButtons/crownButton';
 
 export default {
   components: { CrownButton },
