+++ src/components/nodes/association/association.vue
@@ -23,7 +23,7 @@ import linkConfig from '@/mixins/linkConfig';
 import get from 'lodash/get';
 import associationHead from '@/assets/association-head.svg';
 import { direction } from './associationConfig';
-import CrownConfig from '@/components/crownConfig';
+import CrownConfig from '@/components/crown/crownConfig';
 
 export default {
   components: {
